Use index of previous selection to preserve chosen body type
This feels more future-proof to me, and should effectively keep the same purpose: same model index between selecting/reloading races.

In the future if sex 0 is used to denote a femme body type while 1 is for a masc type, this won't be as intuitive, and we will need to focus on sex instead.